# openapi.yaml
openapi: 3.1.0
info:
  title: machine-launcher
  description: ''
  license:
    name: ''
  version: 0.1.0
paths:
  /api/auth/nonce:
    get:
      tags:
      - handlers_config
      operationId: issue_nonce
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NonceResponse'
  /api/oidc-config:
    get:
      tags:
      - handlers_config
      operationId: oidc_config
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OidcConfigResponse'
  /api/servers:
    get:
      tags:
      - handlers_app
      operationId: machine_status
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Server'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
  /api/servers/start:
    put:
      tags:
      - handlers_app
      operationId: start_machine
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServerName'
        required: true
      responses:
        '202':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
  /api/servers/stop:
    put:
      tags:
      - handlers_app
      operationId: stop_machine
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServerName'
        required: true
      responses:
        '202':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Server'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
components:
  schemas:
    ErrorMessage:
      type: object
      required:
      - error
      properties:
        error:
          type: string
    NonceResponse:
      type: object
      required:
      - nonce
      properties:
        nonce:
          type: string
    OidcConfigResponse:
      type: object
      required:
      - client_id
      - authorization_endpoint
      - token_endpoint
      properties:
        authorization_endpoint:
          type: string
        client_id:
          type: string
        token_endpoint:
          type: string
    Server:
      type: object
      required:
      - name
      - hostname
      - running
      properties:
        hostname:
          type: string
        name:
          type: string
        reason:
          type:
          - string
          - 'null'
        running:
          type: boolean
    ServerName:
      type: object
      required:
      - name
      properties:
        name:
          type: string